fit_board/requirements.txt
asgiref==3.8.1
Django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
